refactor: 엔티티클래스 패키지 모아주기
