+++ src/main/java/persistence/model/Department.java
@@ -1,0 +1,21 @@
+package persistence.model;
+
+import jakarta.persistence.*;
+
+import java.util.List;
+
+@Entity
+public class Department {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @OneToMany
+    @JoinColumn(name = "department_id")
+    private List<Employee> employees;
+
+    // getter, setter, constructors, 등등
+}
+
