+++ src/main/java/persistence/model/Order.java
@@ -1,4 +1,4 @@
-package persistence;
+package persistence.model;
 
 import jakarta.persistence.*;
