+++ src/test/java/persistence/entity/CollectionLoaderTest.java
@@ -2,7 +2,7 @@ package persistence.entity;
 
 import org.junit.jupiter.api.*;
 import persistence.H2DBTestSupport;
-import persistence.OrderLazy;
+import persistence.model.OrderLazy;
 import persistence.entity.collection.CollectionLoader;
 import persistence.sql.mapping.Associations;
