+++ src/test/java/persistence/entity/EntityLoaderAssociationTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.H2DBTestSupport;
-import persistence.Order;
-import persistence.OrderLazy;
-import persistence.OrderLazySet;
+import persistence.model.Order;
+import persistence.model.OrderLazy;
+import persistence.model.OrderLazySet;
 import persistence.entity.collection.PersistentList;
 import persistence.entity.collection.PersistentSet;
 import persistence.sql.mapping.PersistentClass;
