+++ src/test/java/persistence/entity/PersistentCollectionTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.H2DBTestSupport;
-import persistence.OrderItem;
-import persistence.OrderLazy;
+import persistence.model.OrderItem;
+import persistence.model.OrderLazy;
 import persistence.entity.collection.CollectionLoader;
 import persistence.entity.collection.PersistentList;
 import persistence.sql.mapping.Associations;
