+++ src/test/java/persistence/sql/dml/SelectQueryBuilderTest.java
@@ -2,14 +2,10 @@ package persistence.sql.dml;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.Order;
-import persistence.Person;
-import persistence.sql.mapping.Associations;
-import persistence.sql.mapping.Columns;
+import persistence.model.Order;
+import persistence.model.Person;
 import persistence.sql.mapping.PersistentClass;
-import persistence.sql.mapping.TableData;
 
-import static org.assertj.core.api.Assertions.as;
 import static org.assertj.core.api.Assertions.assertThat;
 import static persistence.sql.dml.BooleanExpression.eq;
